requirements.txt
# 실행코드 pip install -r requirements.txt

fastapi
uvicorn
Pillow
torch
python-multipart
ultralytics
aiofiles
requests